requirements.txt
langchain>=1.0.0
langchain-classic>=1.0.0
langchain-openai>=1.0.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
faiss-cpu>=1.8.0
openai>=1.40.0
pypdf>=4.3.1
tiktoken>=0.7.0
streamlit>=1.38.0
python-dotenv>=1.0.1
rank_bm25>=0.2.2
flashrank>=0.2.0
pymupdf>=1.24.0
Pillow>=10.0.0
